Update License
Someone opened a PR to do this a few days ago. I posted a comment. When I was ready to return to the ticket, I couldn't find it as opened or closed. It seemed like a reasonable idea. I will submit this PR for now and sit on it for about a week to give the original author a chance to resubmit.